[flutter_tools] split features into 2 libraries, migrate to null safety
